move singleton to yjs package so other repos (lexical) can use it.
